+++ NEWS.md
@@ -1,3 +1,8 @@
+Dec 31st, 2018
+==============
+* Version `0.35.1` of the server extension and `0.26.1` of the client
+    * Now the bulk geocoder returns `Point` geometries instead of MultiPolygons (#538)
+
 Nov 30th, 2018
 ==============
 * Version `0.35.0` of the server extension